requirements.txt
asgiref==3.2.3
astroid==2.2.5
colorama==0.4.1
cycler==0.10.0
DateTime==4.3
Django==3.0.3
et-xmlfile==1.0.1
glob3==0.0.1
image==1.5.28
imutils==0.5.3
isort==4.3.21
jdcal==1.4.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
matplotlib==3.1.2
mccabe==0.6.1
numpy==1.17.0
opencv-python==4.1.1.26
openpyxl==3.0.3
Pillow==6.2.1
pyipm==0.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
scipy==1.3.1
six==1.12.0
sqlparse==0.3.0
theano==1.0.4
typed-ast==1.4.0
virtualenv==16.7.4
wrapt==1.11.2
zope.interface==4.7.1
